requirements.txt
tensorflow
pytz
pandas
numpy
python-multipart
fastapi
jinja2
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
python-dotenv
google-generativeai
google-ai-generativelanguage
docarray
ipywidgets
jupyter